Update cmc.js
the "initialize" is conflicting with the callback function of the Google Maps API ("http://maps.googleapis.com/maps/api/js?key=YOUR_KEY&sensor=false&callback=initialize") 

When the form is hidden as the user does not have to retype what was wrong. The form should be displayed and the image of "loading" hidden again.